completed all

+++ src/main/java/Packageapp/services/ParcService.java
@@ -1,6 +1,8 @@
 package Packageapp.services;
 
 import Packageapp.exceptions.NotFoundException;
+import Packageapp.models.Geoloc;
+import Packageapp.repositories.GeolocRepository;
 import Packageapp.repositories.ParcRepository;
 import Packageapp.exceptions.DBException;
 import Packageapp.models.Parc;
@@ -15,6 +17,7 @@ import java.util.List;
 public class ParcService {
 
     private final ParcRepository parcRepository;
+    private final GeolocRepository geolocRepository;
 
     public List<Parc> getAllParcs() {
         return this.parcRepository.findAll();
@@ -25,8 +28,10 @@ public class ParcService {
 
         if (parc.getID() != null){
             existing = this.parcRepository.findById(parc.getID()).orElse(null);
-            if (existing == null) throw new NotFoundException("Could not find Geoloc with id : " + parc.getID());
+            if (existing == null) throw new NotFoundException("Could not find Parc with id : " + parc.getID());
         } else {
+            if (parc.getNbtot() < parc.getNbdispo()) throw new DBException("Nombre de places dispos superieur au nombre total");
+
             existing = new Parc();
         }
 
@@ -43,13 +48,29 @@ public class ParcService {
         }
     }
 
+    public Parc getParcById(Long id) throws NotFoundException {
+        return parcRepository.findById(id).orElseThrow(() -> new NotFoundException("Parc not found"));
+    }
+
     public void deleteParc(Long id) throws NotFoundException, DBException {
         Parc existing = this.parcRepository.findById(id).orElse(null);
         if (existing == null) throw new NotFoundException("Can't find parc with id :" + id);
-        try {
-            this.parcRepository.delete(existing);
-        } catch (Exception e) {
+
+        //recup la geoloc liée à la localisation
+        Geoloc geoloc = this.geolocRepository.findById(existing.getGeoloc_ID()).orElse(null);
+        if (geoloc == null) throw new NotFoundException("Can't find geoloc with id :" + existing.getGeoloc_ID());
+
+        try{
+            this.geolocRepository.delete(geoloc);
+            try {
+                this.parcRepository.delete(existing);
+            } catch (Exception e) {
+                throw new DBException("Error with DB");
+            }
+        }catch (Exception e) {
             throw new DBException("Error with DB");
         }
+
+
     }
 }
